Refactor storage package
