+++ src/main/java/tecna/Tecna.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 
 import org.json.simple.parser.ParseException;
 
+import tecna.storage.Storage;
 import tecna.ui.Ui;
 
 /**
